Move hive util to its own file
